Move sequence file and rename filter function
Documenting the filter_covid_genome_metadata function made it
clear that its name should better align with the CladeTime
vernacular ('sequence metadata'). Additionally, it makes more
sense to move sequence.py out of the utility folder, since
sequence-related functions are integral to the package.

util/sequence.py remains in the code base for now, so we can
import the filtering function with its old name for backward
compatibility.
