--- docs/reference/filter_covid_genome_metadata.rst
@@ -1,6 +1,0 @@
-====================================================
-cladetime.util.sequence.filter_covid_genome_metadata
-====================================================
-
-.. autofunction:: cladetime.util.sequence.filter_covid_genome_metadata
-
