+++ docs/reference/index.rst
@@ -4,6 +4,6 @@ API Reference
 .. toctree::
 
    cladetime
-   filter_covid_genome_metadata
+   sequence
    types
