+++ docs/reference/sequence.rst
@@ -1,0 +1,6 @@
+=========
+sequence
+=========
+
+.. autofunction:: cladetime.sequence.filter_sequence_metadata
+
