+++ docs/reference/types.rst
@@ -1,5 +1,5 @@
 =====
-Types
+types
 =====
